refactor: make sdm monitor observer generic
